refactor __init__.py hierarchy of imports
