+++ src/nisarqa/products/__init__.py
@@ -1,0 +1,3 @@
+# keep individual products in their own namespace
+from . import caltools, gcov, gslc, igram, offsets, rslc
+from .product_reader import *
