+++ src/nisarqa/stats_h5_globals.py
